chore: convert most inventory calls to transfer api

+++ src/main/java/dev/galacticraft/mod/util/DrawableUtil.java
@@ -49,17 +49,17 @@ public class DrawableUtil {
         textRenderer.draw(matrices, text.asOrderedText(), (float) (x - textRenderer.getWidth(text) / 2), (float) y, color);
     }
 
-    public static void drawOxygenBuffer(MatrixStack matrices, int x, int y, int oxygen, int capacity) {
+    public static void drawOxygenBuffer(MatrixStack matrices, int x, int y, long oxygen, long capacity) {
         if (oxygen == 0 && capacity == 0) capacity = 1;
-        drawOxygenBuffer(matrices, x, y, (float) oxygen / (float) capacity);
+        drawOxygenBuffer(matrices, x, y, (double) oxygen / (double) capacity);
     }
 
-    public static void drawOxygenBuffer(MatrixStack matrices, int x, int y, float scale) {
+    public static void drawOxygenBuffer(MatrixStack matrices, int x, int y, double scale) {
         RenderSystem.setShader(GameRenderer::getPositionTexShader);
         RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
         RenderSystem.setShaderTexture(0, Constant.ScreenTexture.OVERLAY);
         drawProgressTexture(matrices, x, y, 0, Constant.TextureCoordinate.OXYGEN_DARK_X, Constant.TextureCoordinate.OXYGEN_DARK_Y, Constant.TextureCoordinate.OVERLAY_WIDTH, Constant.TextureCoordinate.OVERLAY_HEIGHT, 128, 128);
-        drawProgressTexture(matrices, x, (int) (y + Constant.TextureCoordinate.OVERLAY_HEIGHT - (Constant.TextureCoordinate.OVERLAY_HEIGHT * scale)), 0, Constant.TextureCoordinate.OXYGEN_LIGHT_X, Constant.TextureCoordinate.OXYGEN_LIGHT_Y, Constant.TextureCoordinate.OVERLAY_WIDTH, Constant.TextureCoordinate.OVERLAY_HEIGHT * scale, 128, 128);
+        drawProgressTexture(matrices, x, (int) (y + Constant.TextureCoordinate.OVERLAY_HEIGHT - (Constant.TextureCoordinate.OVERLAY_HEIGHT * scale)), 0, Constant.TextureCoordinate.OXYGEN_LIGHT_X, Constant.TextureCoordinate.OXYGEN_LIGHT_Y, Constant.TextureCoordinate.OVERLAY_WIDTH, (float) (Constant.TextureCoordinate.OVERLAY_HEIGHT * scale), 128, 128);
     }
 
     public static boolean isWithin(double mouseX, double mouseY, int x, int y, int width, int height) {
